feat: 添加权限验证

+++ back-end/src/api/tag.rs
@@ -1,5 +1,7 @@
 use crate::{repository::mongodb_repos::MongoRepo};
 use crate::{api::response::MyResponse};
+use crate::{api::auth::auth_error};
+use crate::api::user_data::UserData;
 use actix_web::{
   post,
   get,
@@ -17,23 +19,27 @@ pub struct Tag_Info {
 }
 
 #[get("/add_tag")]
-pub async fn add_tag(db: Data<MongoRepo>, info: Query<Tag_Info>) -> HttpResponse  {
-  let id: String = info.name.clone();
-  if id.is_empty() {
-    return HttpResponse::BadRequest().body("invalid ID");
-  };
-  let result = db.add_tag(&id, &info.color).await;
-  match result {
-    Ok(()) => {
-      let result: MyResponse<String> = MyResponse {
-        result: "0".to_string(),
-        message: "add done".to_string(),
-        data: None
-      };
-  
-      HttpResponse::Ok().json(result)
-    },
-    Err(err) => HttpResponse::InternalServerError().body(err.to_string()),
+pub async fn add_tag(db: Data<MongoRepo>, user: Option<UserData>, info: Query<Tag_Info>) -> HttpResponse  {
+  if let Some(user) = user {
+    let id: String = info.name.clone();
+    if id.is_empty() {
+      return HttpResponse::BadRequest().body("invalid ID");
+    };
+    let result = db.add_tag(&id, &info.color).await;
+    match result {
+      Ok(()) => {
+        let result: MyResponse<String> = MyResponse {
+          result: "0".to_string(),
+          message: "add done".to_string(),
+          data: None
+        };
+    
+        HttpResponse::Ok().json(result)
+      },
+      Err(err) => HttpResponse::InternalServerError().body(err.to_string()),
+    }
+  } else {
+    auth_error()
   }
 }
 
@@ -45,30 +51,34 @@ pub struct Update_Tag_Info {
 }
 
 #[get("/update_tag")]
-pub async fn update_tag(db: Data<MongoRepo>, info: Query<Update_Tag_Info>) -> HttpResponse  {
-  let id: String = info.id.clone();
-  let update_result = db.edit_tag(&info.id, &info.name, &info.color).await;
-  match update_result {
-    Ok(update) => {
-        if update.matched_count == 1 {
-            let updated_tag_info = db.get_photo(&id).await;
-            return match updated_tag_info {
-              Ok(tag) => {
-                let result = MyResponse {
-                  result: "0".to_string(),
-                  message: "update done".to_string(),
-                  data: Some(tag)
-                };
-            
-                return HttpResponse::Ok().json(result)
-              },
-              Err(err) => HttpResponse::InternalServerError().body(err.to_string()),
-            };
-        } else {
-            return HttpResponse::NotFound().body("No tag found with specified ID");
-        }
+pub async fn update_tag(db: Data<MongoRepo>, user: Option<UserData>, info: Query<Update_Tag_Info>) -> HttpResponse  {
+  if let Some(user) = user {
+    let id: String = info.id.clone();
+    let update_result = db.edit_tag(&info.id, &info.name, &info.color).await;
+    match update_result {
+      Ok(update) => {
+          if update.matched_count == 1 {
+              let updated_tag_info = db.get_photo(&id).await;
+              return match updated_tag_info {
+                Ok(tag) => {
+                  let result = MyResponse {
+                    result: "0".to_string(),
+                    message: "update done".to_string(),
+                    data: Some(tag)
+                  };
+              
+                  return HttpResponse::Ok().json(result)
+                },
+                Err(err) => HttpResponse::InternalServerError().body(err.to_string()),
+              };
+          } else {
+              return HttpResponse::NotFound().body("No tag found with specified ID");
+          }
+      }
+      Err(err) => HttpResponse::InternalServerError().body(err.to_string()),
     }
-    Err(err) => HttpResponse::InternalServerError().body(err.to_string()),
+  } else {
+    auth_error()
   }
 }
 
@@ -79,26 +89,30 @@ pub struct Delete_Info {
 
 
 #[get("/delete_tag")]
-pub async fn delte_tag(db: Data<MongoRepo>, info: Query<Delete_Info>) -> HttpResponse  {
-  let id: String = info.id.clone();
-  if id.is_empty() {
-    return HttpResponse::BadRequest().body("invalid ID");
-  };
-  let result = db.delete_tag(&id).await;
-  match result {
-      Ok(res) => {
-          if res.deleted_count == 1 {
-            let result: MyResponse<String> = MyResponse {
-              result: "0".to_string(),
-              message: "tag successfully deleted!".to_string(),
-              data: None
-            };
-        
-            return HttpResponse::Ok().json(result);
-          } else {
-              return HttpResponse::NotFound().json("tag with specified ID not found!");
-          }
-      }
-      Err(err) => HttpResponse::InternalServerError().body(err.to_string()),
+pub async fn delte_tag(db: Data<MongoRepo>, user: Option<UserData>, info: Query<Delete_Info>) -> HttpResponse  {
+  if let Some(user) = user {
+    let id: String = info.id.clone();
+    if id.is_empty() {
+      return HttpResponse::BadRequest().body("invalid ID");
+    };
+    let result = db.delete_tag(&id).await;
+    match result {
+        Ok(res) => {
+            if res.deleted_count == 1 {
+              let result: MyResponse<String> = MyResponse {
+                result: "0".to_string(),
+                message: "tag successfully deleted!".to_string(),
+                data: None
+              };
+          
+              return HttpResponse::Ok().json(result);
+            } else {
+                return HttpResponse::NotFound().json("tag with specified ID not found!");
+            }
+        }
+        Err(err) => HttpResponse::InternalServerError().body(err.to_string()),
+    }
+  } else {
+    auth_error()
   }
 }
